Subclassed zevatech auto feeder so that we don't modify the base impl.
Added a getsDoesFeederDoPick() func to the feeder interface, as our ZevatechFeeder implements the pick functionality inside the feed() call - this is done so we can check the vacuum and keep retrying the picks until successful. All the default feeders do not set the doesFeederDoPick() bool, so functionality remains the same.